chore(react): review and corrections
