+++ packages/react/src/components/AppWrapper/index.tsx
@@ -1,16 +1,12 @@
 import { Outlet } from "react-router-dom";
 import { DialogManagerProvider, DrawerProvider } from "../../providers";
+import type { HeaderOptionType } from "../../types";
 import { BaseWrapper } from "../BaseWrapper";
 import { ActionDialog } from "./ActionDialog";
 import { AppHeader } from "./AppHeader";
 import { SidebarDrawer } from "./Drawer";
 
-export type OptionType = {
-  path: string;
-  label: string;
-};
-
-const OPTIONS: OptionType[] = [
+const OPTIONS: HeaderOptionType[] = [
   {
     path: "/collections",
     label: "Collections",
