+++ packages/react/src/components/PluginCard.tsx
@@ -16,11 +16,13 @@ export function PluginCard(props: PluginCard) {
   );
 
   return (
-    <div className="flex items-center gap-2 md:gap-2.5 lg:gap-3 rounded md:rounded-md bg-secondary-100 hover:shadow-lg p-3 md:p-3.5 lg:p-4 transition-all ease-in-out dark:hover:shadow-none dark:bg-secondary-900 dark:hover:bg-secondary-800">
+    <div className="flex items-center gap-1.5 md:gap-2 rounded md:rounded-md bg-secondary-100 hover:shadow-lg p-3 md:p-3.5 lg:p-4 transition-all ease-in-out dark:hover:shadow-none dark:bg-secondary-900 dark:hover:bg-secondary-800">
       <Suspense fallback="Loading...">
-        <Icon className="size-4 stroke-2" />
+        <Icon className="size-4 md:size-5 stroke-2" />
       </Suspense>
-      {props.label}
+      <h3 className="text-sm md:text-base font-medium leading-tight md:leading-none lg:leading-none">
+        {props.label}
+      </h3>
     </div>
   );
 }
