+++ packages/react/src/types/header-option.ts
@@ -1,0 +1,4 @@
+export type HeaderOptionType = {
+  path: string;
+  label: string;
+};
