+++ packages/react/src/types/index.ts
@@ -1,2 +1,3 @@
 export type { CollectionType } from "./collection";
+export type { HeaderOptionType } from "./header-option";
 export type { PluginType } from "./plugin";
